# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel Order API",
        "description": "Api de \"gestão\" de pedido de viagem",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/orders/{orderId}/approve": {
            "patch": {
                "tags": [
                    "Orders"
                ],
                "summary": "Order approval",
                "description": "Approve an order",
                "operationId": "approve-order",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "ID of the order to approve",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "01F8MECHZX3TBDSZ7XRADM79XE"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order approved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order approved"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": [
                                                "Order already approved",
                                                "Approved order cannot be canceled"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{orderId}/cancel": {
            "patch": {
                "tags": [
                    "Orders"
                ],
                "summary": "Order cancellation",
                "description": "Cancel an order",
                "operationId": "cancel-order",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "ID of the order to canceled",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "01F8MECHZX3TBDSZ7XRADM79XE"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order canceled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order canceled"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": [
                                                "Order already canceled",
                                                "Canceled order cannot be approved"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List all orders (if user has admin role, can list from all users omitting userId query parameter)",
                "description": "List all orders",
                "operationId": "list-order",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "requested",
                            "enum": [
                                "requested",
                                "approved",
                                "canceled"
                            ]
                        }
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "01kmknec99e43swm1r86yryt38"
                        }
                    },
                    {
                        "name": "destination",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "São Paulo - SP"
                        }
                    },
                    {
                        "name": "departureDate",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "2026-10-21"
                        }
                    },
                    {
                        "name": "arrivalDate",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "2026-10-27"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orders list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderCollectionResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Order creation",
                "description": "Create a new order",
                "operationId": "create-order",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CreateOrderRequest": {
                "required": [
                    "destination",
                    "departureDate",
                    "arrivalDate"
                ],
                "properties": {
                    "destination": {
                        "description": "Destination of the order",
                        "type": "string",
                        "example": "São Paulo - SP"
                    },
                    "departureDate": {
                        "description": "Departure date of the order",
                        "type": "string",
                        "example": "2026-05-21"
                    },
                    "arrivalDate": {
                        "description": "Arrival date of the order",
                        "type": "string",
                        "example": "2026-05-27"
                    }
                },
                "type": "object"
            },
            "OrderCollectionResource": {
                "required": [
                    "data",
                    "meta"
                ],
                "properties": {
                    "data": {
                        "description": "List of orders",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderResource"
                        }
                    },
                    "meta": {
                        "description": "Pagination metadata",
                        "properties": {
                            "total": {
                                "type": "integer",
                                "example": 100
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 10
                            },
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "OrderResource": {
                "required": [
                    "id",
                    "userId",
                    "userName",
                    "destination",
                    "departureDate",
                    "arrivalDate",
                    "status",
                    "createdAt",
                    "updatedAt"
                ],
                "properties": {
                    "id": {
                        "description": "Order Id",
                        "type": "string",
                        "example": "01KMKXT8169DZBFAY64958180J"
                    },
                    "userId": {
                        "description": "User Id",
                        "type": "string",
                        "example": "01kmknec99e43swm1r86yryt38"
                    },
                    "userName": {
                        "description": "User name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "destination": {
                        "description": "Travel destination",
                        "type": "string",
                        "example": "São Paulo - SP"
                    },
                    "departureDate": {
                        "description": "Departure date",
                        "type": "string",
                        "example": "2026-05-21"
                    },
                    "arrivalDate": {
                        "description": "Arrival date",
                        "type": "string",
                        "example": "2026-05-28"
                    },
                    "status": {
                        "description": "Order status",
                        "type": "string",
                        "example": "approved",
                        "enum": [
                            "approved",
                            "approved",
                            "cancelled"
                        ]
                    },
                    "createdAt": {
                        "description": "Created at",
                        "type": "string",
                        "example": "2026-05-28 01:25:12"
                    },
                    "updatedAt": {
                        "description": "Updated at",
                        "type": "string",
                        "example": "2026-05-28 01:27:01"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}